Add zero allocation version of toRfc3339
The two allocations in the original implementation were frankly
unacceptable for any performance sensitive code.
The old version is intact and has been renamed to `toRfc3339Alloc` and a
new version has been added that allows a library consumer to pass in a
buffer that is checked to be large enough. This buffer could be on the
stack or the heap as demonstrated by the examples in the readme.